requirements.txt
fastapi>=0.116.1,<1.0.0
httpx>=0.28.1,<1.0.0
markdown>=3.8.2,<4.0.0
orjson>=3.11.3,<4.0.0
playwright>=1.55.0,<2.0.0
pydantic-settings>=2.10.1,<3.0.0
pytest>=8.4.2,<9.0.0
pytest-asyncio>=1.2.0,<2.0.0
redis>=6.4.0,<7.0.0
sse-starlette>=2.4.1,<3.0.0
uvicorn[standard]>=0.35.0,<1.0.0
